move cheetah_cmd into distribution
