+++ ChangeLog
@@ -1,9 +1,18 @@
+Tue Oct 16 2018 Christopher O'Brien <obriencj@gmail.com> - 1.5.0-0
+	Python 3 support (thanks Konstantin Shemyak)
+
+	Moved to Cheetah3
+
+	check text file changes
+
+	switch from optparse to argpase
+
 Mon Apr 10 2017 Christopher O'Brien <obriencj@gmail.com> - 1.4.0-0
 	Python 2.7+
 
 	Support for signing and verifying JAR signatures (adds
 	M2Crypto dependency)
-	
+
 	fixed issues with auto-exploding a dist and cleaning up
 	afterwards
 
@@ -16,7 +25,7 @@ Mon Apr 10 2017 Christopher O'Brien <obriencj@gmail.com> - 1.4.0-0
 	more crypto hashes supported in manifests
 
 	removed PyXML dependency
-	
+
 	Use entry_points for CLI invocation
 
 	manifest enhancements
