+++ javatools/cheetah/__init__.py
@@ -21,7 +21,7 @@ author: Christopher O'Brien <obriencj@gmail.com>
 license: LGPL
 """
 
-from Cheetah.Template import Template
+
 from pkgutil import iter_modules
 
 
@@ -33,6 +33,7 @@ def _iter_templates():
     # pylint: disable=W0406
     # needed for introspection
     import javatools.cheetah
+    from Cheetah.Template import Template
 
     for _, name, _ in iter_modules(__path__):
         __import__("javatools.cheetah." + name)
